Move NXlog tests to file
