+++ tests/nexus_test.py
@@ -99,112 +99,6 @@ def test_nx_class_can_be_bytes(h5root):
     assert group.nx_class == NXlog
 
 
-def test_nxobject_log(h5root):
-    da = sc.DataArray(sc.array(dims=['time'], values=[1.1, 2.2, 3.3]),
-                      coords={
-                          'time':
-                          sc.epoch(unit='ns') +
-                          sc.array(dims=['time'], unit='s', values=[4.4, 5.5, 6.6]).to(
-                              unit='ns', dtype='int64')
-                      })
-    log = snx.create_class(h5root, 'log', NXlog)
-    snx.create_field(log, 'value', da.data)
-    snx.create_field(log, 'time', da.coords['time'] - sc.epoch(unit='ns'))
-    log = snx.Group(log, definitions=snx.base_definitions())
-    assert sc.identical(log[...], da)
-
-
-def test_nxlog_with_missing_value_triggers_fallback(nxroot):
-    time = sc.epoch(unit='ns') + sc.array(
-        dims=['time'], unit='s', values=[4.4, 5.5, 6.6]).to(unit='ns', dtype='int64')
-    log = nxroot['entry'].create_class('log', NXlog)
-    log['time'] = time - sc.epoch(unit='ns')
-    loaded = log[()]
-    # Fallback to DataGroup, but we still have partial info from NXlog: dim is time
-    assert_identical(loaded, sc.DataGroup(time=time))
-
-
-def test_nxlog_length_1(h5root):
-    nxroot = snx.Group(h5root, definitions=snx.base_definitions())
-    da = sc.DataArray(
-        sc.array(dims=['time'], values=[1.1]),
-        coords={
-            'time':
-            sc.epoch(unit='ns') +
-            sc.array(dims=['time'], unit='s', values=[4.4]).to(unit='ns', dtype='int64')
-        })
-    log = nxroot.create_class('log', NXlog)
-    log['value'] = da.data
-    log['time'] = da.coords['time'] - sc.epoch(unit='ns')
-    assert sc.identical(log[...], da)
-
-
-def test_nxlog_length_1_two_dims_no_time_defaults_inner_dim_name(nxroot):
-    var = sc.array(dims=['time', 'ignored'], values=[[1.1]])
-    log = nxroot['entry'].create_class('log', NXlog)
-    log['value'] = var
-    assert_identical(log[...], sc.DataArray(var.rename(ignored='dim_1')))
-
-
-def test_nxlog_length_1_two_dims_with_time_defaults_inner_dim_name(nxroot):
-    da = sc.DataArray(
-        sc.array(dims=['time', 'ignored'], values=[[1.1]]),
-        coords={
-            'time':
-            sc.epoch(unit='ns') +
-            sc.array(dims=['time'], unit='s', values=[4.4]).to(unit='ns', dtype='int64')
-        })
-    log = nxroot['entry'].create_class('log', NXlog)
-    log['value'] = da.data
-    log['time'] = da.coords['time'] - sc.epoch(unit='ns')
-    assert sc.identical(log[...], da.rename(ignored='dim_1'))
-
-
-def test_nxlog_axes_replaces_time_dim(nxroot):
-    da = sc.DataArray(
-        sc.array(dims=['time', 'ignored'], values=[[1.1]]),
-        coords={
-            'time':
-            sc.epoch(unit='ns') +
-            sc.array(dims=['time'], unit='s', values=[4.4]).to(unit='ns', dtype='int64')
-        })
-    log = nxroot['entry'].create_class('log', NXlog)
-    log._group.attrs['axes'] = ['yy', 'xx']
-    log['value'] = da.data
-    log['time'] = da.coords['time'] - sc.epoch(unit='ns')
-    expected = sc.DataArray(sc.array(dims=['yy', 'xx'], values=[[1.1]]),
-                            coords={'time': da.coords['time'].squeeze()})
-    assert sc.identical(log[...], expected)
-
-
-def test_nxlog_three_dims_with_time_of_length_1(nxroot):
-    da = sc.DataArray(
-        sc.array(dims=['time', 'a', 'b'], values=np.arange(9.).reshape(1, 3, 3)),
-        coords={
-            'time':
-            sc.epoch(unit='ns') +
-            sc.array(dims=['time'], unit='s', values=[4.4]).to(unit='ns', dtype='int64')
-        })
-    log = nxroot['entry'].create_class('log', NXlog)
-    log['value'] = da.data
-    log['time'] = da.coords['time'] - sc.epoch(unit='ns')
-    loaded = log[...]
-    assert_identical(
-        loaded.data,
-        sc.array(dims=['time', 'dim_1', 'dim_2'], values=np.arange(9.).reshape(1, 3,
-                                                                               3)))
-
-
-def test_nxlog_with_shape_0(nxroot):
-    da = sc.DataArray(sc.ones(dims=['time', 'ignored'], shape=(0, 1)),
-                      coords={'time': sc.ones(dims=['time'], shape=(0, ), unit='s')})
-    log = nxroot['entry'].create_class('log', NXlog)
-    log['value'] = da.data
-    log['time'] = da.coords['time']
-    da.coords['time'] = sc.datetimes(dims=['time'], values=[], unit='ns')
-    assert_identical(log[...], da.rename(ignored='dim_1'))
-
-
 def test_nxobject_event_data(nxroot):
     event_data = nxroot['entry'].create_class('events_0', NXevent_data)
     assert event_data.nx_class == NXevent_data
